+++ app/Filament/Resources/Sessions/SessionResource.php
@@ -1,0 +1,54 @@
+<?php
+
+namespace App\Filament\Resources\Sessions;
+
+use App\Filament\Resources\Sessions\Pages\ListSessions;
+use App\Filament\Resources\Sessions\Tables\SessionsTable;
+use App\Models\Session;
+use BackedEnum;
+use Filament\Resources\Resource;
+use Filament\Schemas\Schema;
+use Filament\Support\Icons\Heroicon;
+use Filament\Tables\Table;
+use Illuminate\Database\Eloquent\Model;
+
+class SessionResource extends Resource
+{
+    protected static ?string $model = Session::class;
+
+    protected static string|BackedEnum|null $navigationIcon = Heroicon::OutlinedCpuChip;
+
+    public static function canCreate(): bool
+    {
+        return false;
+    }
+
+    public static function canEdit(Model $record): bool
+    {
+        return false;
+    }
+
+    public static function form(Schema $schema): Schema
+    {
+        return $schema;
+    }
+
+    public static function table(Table $table): Table
+    {
+        return SessionsTable::configure($table);
+    }
+
+    public static function getRelations(): array
+    {
+        return [
+            //
+        ];
+    }
+
+    public static function getPages(): array
+    {
+        return [
+            'index' => ListSessions::route('/'),
+        ];
+    }
+}